+++ src/plater/execution_mode/threading_mode.rs
@@ -1,0 +1,7 @@
+use crate::plater::solution::Solution;
+
+#[derive(Clone)]
+pub enum ThreadingMode {
+    SingleThreaded,
+    MultiThreaded,
+}
