+++ src/stl/request.rs
@@ -1,11 +1,13 @@
 use std::collections::HashMap;
 
 use crate::plater;
+use crate::plater::execution_mode::multi_threaded_runner::MultiThreadedRunner;
+use crate::plater::execution_mode::single_threaded_runner::SingleThreadedRunner;
+use crate::plater::execution_mode::threading_mode::ThreadingMode;
 use crate::plater::plate_shape::{PlateShape, Shape};
 use crate::plater::progress_config::{FutureKillSwitch, ProgressConfig};
 use crate::plater::request::{
     Algorithm, BedExpansionMode, ConfigOrder, PlacingError, PointEnumerationMode, Strategy,
-    ThreadingMode,
 };
 use crate::plater::solution::Solution;
 use crate::stl::model::Model;
@@ -23,7 +25,10 @@ impl Request {
         &self,
         config: ProgressConfig<T, F1, F2, F3>,
     ) -> Result<T, PlacingError> {
-        self.request.process(config)
+        match &self.request.algorithm.threading_mode {
+            ThreadingMode::SingleThreaded => SingleThreadedRunner::new(&self.request).place(config),
+            ThreadingMode::MultiThreaded => MultiThreadedRunner::new(&self.request).place(config),
+        }
     }
 
     pub fn new(plate_shape: Shape, resolution: f64) -> Self {
